+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add "Hide image" button to images with content warning.
 - Show lightbox when clicking on image attachment.
 - Added alt text to attached images.
+- Added "Repost with comment" to post menu.
 
 ### Changed